Show tmdb movie reviews in review history screen
Tags: feature

+++ app/src/main/kotlin/com/rbraithwaite/untitledmovieapp/core/repositories/MediaRepository.kt
@@ -23,13 +23,14 @@ interface MediaRepository {
 
     suspend fun addOrUpdateTmdbLite(tmdblite: TmdbLite)
 
+    // REFACTOR [24-01-27 12:27p.m.] -- this should be moved to ReviewRepository.
     // REFACTOR [23-12-20 3:41p.m.] -- this is very lazy lol - I need to rethink how I'm adding reviews,
     //  and core data in general, like have one common addMediaReview(<media-specific-id>, review)
     //  ---
     //  re: core data, I need to think about how to handle tmdb data.
     // SMELL [23-12-20 3:43p.m.] -- should this be saying 'tmdb'? is that too much info here?
     suspend fun addTmdbMovieReview(
-        tmdbMovieId: Int,
+        tmdbMovieId: Long,
         review: MediaReview
     )
